move a helper to llm
