+++ alphastats/gui/utils/llm_helper.py
@@ -1,30 +1,9 @@
 from pathlib import Path
-from typing import Dict, List
+from typing import List
 
-import pandas as pd
 import streamlit as st
 
 from alphastats.gui.utils.ui_helper import StateKeys
-
-
-def get_subgroups_for_each_group(
-    metadata: pd.DataFrame,
-) -> Dict:
-    """
-    Get the unique values for each column in the metadata file.
-
-    Args:
-        metadata (pd.DataFrame, optional): The metadata dataframe (which sample has which disease/treatment/condition/etc).
-
-    Returns:
-        dict: A dictionary with the column names as keys and a list of unique values as values.
-    """
-    groups = [str(group) for group in metadata.columns.to_list()]
-    group_to_subgroup_values = {
-        group: [str(subgroup) for subgroup in metadata[group].unique().tolist()]
-        for group in groups
-    }
-    return group_to_subgroup_values
 
 
 def display_proteins(overexpressed: List[str], underexpressed: List[str]) -> None:
